# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: COMMIT API
  version: "1.0.0"
  description: COMMIT 프로젝트의 주요 API 명세
servers:
  - url: http://localhost:8080
paths:
...
  /api/community:
    post:
      tags: [커뮤니티]
      summary: 게시글 작성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
            example:
              title: "첫 번째 게시글"
              content: "안녕하세요!"
      responses:
        '200':
          description: 작성 결과
          content:
            application/json:
              schema:
                type: object
                properties:
                  communityId:
                    type: integer
                  message:
                    type: string
              example:
                communityId: 10
                message: "작성 완료"
    get:
      tags: [커뮤니티]
      summary: 게시글 목록 조회
      parameters:
        - in: query
          name: search
          schema:
            type: string
          example: "운동"
        - in: query
          name: page
          schema:
            type: integer
          example: 1
        - in: query
          name: size
          schema:
            type: integer
          example: 10
      responses:
        '200':
          description: 게시글 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommunityListResponse'
              example:
                - communityId: 10
                  title: "첫 번째 게시글"
                  content: "안녕하세요!"
                  author: "홍길동"
                  createdAt: "2024-06-01T12:00:00Z"
...
components:
  schemas:
...
    CommunityListResponse:
      type: object
      properties:
        communityId:
          type: integer
        title:
          type: string
        content:
          type: string
        author:
          type: string
        createdAt:
          type: string
          format: date-time
...